refactor(callbacks): split register_callbacks — Models + Forecast (Step 10c)
Second wave of the ``register_callbacks`` decomposition (issue #87,
Step 10c). Builds on PR #109 which moved Overview + US Grid callbacks
into their tab modules; this PR moves the two largest remaining
tab-bounded blocks.

## What moves in this PR

### Models tab (3 callbacks)

* ``update_models_title``
* ``update_models_leaderboard``
* ``update_models_tab`` (the 6-output diagnostics callback with v1
  compute fallback and Redis fast path)

All move from ``register_callbacks`` into a new
``register_models_callbacks(app)`` in ``components/_callbacks_models.py``.

The section header was historically labeled
``# ── 4. TAB 1: DEMAND FORECAST ──`` even though every output id
the callbacks write to is a Models / Diagnostics tab id
(``models-title``, ``models-leaderboard``, ``tab3-metrics-table``,
``tab3-residuals-*``, ``tab3-error-heatmap``, ``tab3-shap``).
Renamed to ``# ── 4. MODELS / DIAGNOSTICS TAB ──`` in callbacks.py
to match what it actually drives.

### Forecast tab (9 callbacks)

The entire Forecast surface — drivers / generation / scenarios panel
toggles, inline panel content callbacks, scenario presets, slider
readouts, title + model card, the big 9-output
``update_demand_outlook`` callback, and both Forecast Replay
callbacks (NEXD-14 selector + overlay) — moves into
``register_forecast_callbacks(app)`` in
``components/_callbacks_forecast.py``.

The Forecast callbacks reach into ``_callbacks_overview`` for three
panel builders (``_build_drivers_panel``, ``_build_generation_panel``,
``_build_scenarios_panel``) that originally lived in callbacks.py
but were already extracted to the Overview module in #105. They
belong to the Forecast tab in spirit — imported lazily here to keep
the dependency graph one-way (Forecast → Overview) and avoid a
circular import at module load.

## Test patches updated

* ``TestDemandOutlookV1`` (4 tests) in
  ``tests/unit/test_callbacks_v1_paths.py`` — patches rewired from
  ``components.callbacks.{redis_get,_run_forecast_outlook,_outlook_tab_from_redis,_add_confidence_bands,_add_trailing_actuals}``
  to the ``_callbacks_forecast`` namespace. ``update_demand_outlook``
  now reads those names from its own module.
* ``test_v1_fallback_uses_shared_metrics`` in
  ``test_models_tab_consistency.py`` — introspect target changed
  from ``cb.register_callbacks`` to
  ``cm.register_models_callbacks`` (the moved callback's source now
  lives there).

## Public-import surface preserved

callbacks.py re-imports each callback-supporting helper. ``# noqa: F401
— re-export`` markers added to symbols only the moved callbacks use
inside their now-extracted module bodies (e.g. ``_build_models_leaderboard``,
``_format_metric``, ``_outlook_tab_from_redis``).

Also cleaned up the stale ``# ── HELPER FUNCTIONS ──`` / ``# ── Defensive
readers ──`` comment block at the bottom of callbacks.py — those
sections used to introduce blocks of helpers that have all migrated
to tab modules. Replaced with a brief shim-pattern explanation that
points to the explicit re-import block at the top of the file.

## Size impact

callbacks.py: 2,076 → 1,265 (–811 lines, –39%)
_callbacks_models.py: 256 → 525 (+269)
_callbacks_forecast.py: 786 → 1,406 (+620)

**callbacks.py crosses below 1,300 lines.** Remaining ~1,265 is data
loading + persona switching + URL state + NEXD-* feature callbacks
that are foundational rather than tab-specific.

## Cumulative progress

* PRs #98–#108 (helpers per-tab): 7,167 → 2,255
* PR #109 (Overview + US Grid callbacks): 2,255 → 2,076
* this PR (Models + Forecast callbacks): 2,076 → 1,265

**Total: −5,902 lines from the monolith across 13 PRs (−82.4%).**

## What's left

* **Step 10d** — Generation TAB 4 + Scenario TAB 6 (likely hidden per
  the v2 shell; may be deletion candidates rather than migration
  targets). Both still in callbacks.py.
* **Step 10e** — small Risk / weather-context callbacks if any.
* **Final** — update ``docs/internal/CALLBACKS_DECOMPOSITION_PLAN.md``.

Cross-cutting callbacks (data loading, persona, URL state, NEXD
features, header freshness, smart defaults, cross-tab links,
confidence badges, briefing mode) stay in callbacks.py — they're
foundational rather than tab-specific.

## Tests

All affected test classes pass locally:

* ``TestDemandOutlookV1`` (4 tests, v1_paths) — green with rewired patches
* ``TestLeaderboardTableConsistency`` (3 tests, models_tab_consistency)
  — green; ``test_v1_fallback_uses_shared_metrics`` now introspects
  ``register_models_callbacks``
* ``test_callbacks_v1_paths.py`` (40 tests) — all green
* ``ruff check`` + ``ruff format --check`` clean

Full ``pytest tests/unit`` run still in progress at commit time —
pre-existing test-isolation flake in ``test_sprint4.py`` (3 tests
that read ``sys.modules["components.callbacks"]`` without ensuring
the import happened first) is unaffected.
