+++ components/callbacks.py
@@ -49,26 +49,26 @@ from components._callbacks_backtest import (
     _run_backtest_for_horizon,  # noqa: F401 — re-export (tests/unit/test_callbacks_*)
 )
 from components._callbacks_forecast import (
-    _add_confidence_bands,
-    _add_trailing_actuals,
+    _add_confidence_bands,  # noqa: F401 — re-export (callback now lives in _callbacks_forecast)
+    _add_trailing_actuals,  # noqa: F401 — re-export (callback now lives in _callbacks_forecast)
     _confidence_half_width,  # noqa: F401 — re-export (tests/unit/test_callbacks_helpers.py)
     _create_future_features,  # noqa: F401 — re-export (tests/unit/test_callbacks_helpers.py)
-    _outlook_tab_from_redis,
-    _run_forecast_outlook,
+    _outlook_tab_from_redis,  # noqa: F401 — re-export (callback now lives in _callbacks_forecast)
+    _run_forecast_outlook,  # noqa: F401 — re-export (callback now lives in _callbacks_forecast)
+    register_forecast_callbacks,
 )
 from components._callbacks_generation import (
     _generation_tab_from_redis,  # noqa: F401 — re-export (tests/unit/test_redis_fast_paths.py); fast path is currently orphaned in register_callbacks
 )
 from components._callbacks_models import (
-    _format_metric,
-    _get_feature_importance,
-    _models_tab_from_redis,
+    _format_metric,  # noqa: F401 — re-export (callback now lives in _callbacks_models)
+    _get_feature_importance,  # noqa: F401 — re-export (callback now lives in _callbacks_models)
+    _models_tab_from_redis,  # noqa: F401 — re-export (callback now lives in _callbacks_models)
+    register_models_callbacks,
 )
 from components._callbacks_overview import (
     _build_changes_card,  # noqa: F401 — re-export (tests + helper-callable surface)
-    _build_drivers_panel,
-    _build_generation_panel,
-    _build_models_leaderboard,
+    _build_models_leaderboard,  # noqa: F401 — re-export (tests/unit/test_models_tab_consistency.py)
     _build_overview_briefing,  # noqa: F401 — re-export (tests + helper-callable surface)
     _build_overview_data_health,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
     _build_overview_digest,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
@@ -82,7 +82,6 @@ from components._callbacks_overview import (
     _build_overview_title,  # noqa: F401 — re-export (callback now lives in _callbacks_overview)
     _build_persona_kpis,  # noqa: F401 — re-export (tests/unit/test_callbacks_helpers.py)
     _build_risk_insight,
-    _build_scenarios_panel,
     _build_weather_context,
     _fetch_generation_cached,  # noqa: F401 — re-export (tests/unit/test_callbacks_*)
     _spotlight_model_accuracy,  # noqa: F401 — re-export (tests/unit/test_tab_overview.py)
@@ -118,10 +117,8 @@ from components._callbacks_shared import (
 from components._callbacks_weather import (
     _weather_tab_from_redis,  # noqa: F401 — re-export (tests/unit/test_redis_fast_paths.py); fast path is currently orphaned in register_callbacks
 )
-from components.accessibility import LINE_STYLES
 from components.cards import (
     build_alert_card,
-    build_model_metrics_card,
     build_page_title,
 )
 from config import (
@@ -525,247 +522,12 @@ def register_callbacks(app):
 
         return snapshots, _json.dumps(changes)
 
-    # ── 4. TAB 1: DEMAND FORECAST ─────────────────────────────
-
-    @app.callback(
-        Output("models-title", "children"),
-        [
-            Input("region-selector", "value"),
-            Input("dashboard-tabs", "active_tab"),
-        ],
-    )
-    def update_models_title(region, active_tab):
-        """Page title for the Models tab."""
-        if active_tab != "tab-models":
-            return no_update
-
-        region = region or "FPL"
-        region_name = REGION_NAMES.get(region, region)
-        return build_page_title(
-            "Models",
-            f"Forecast accuracy, residuals, and feature importance · {region_name}",
-        )
-
-    @app.callback(
-        Output("models-leaderboard", "children"),
-        [
-            Input("region-selector", "value"),
-            Input("dashboard-tabs", "active_tab"),
-        ],
-    )
-    def update_models_leaderboard(region, active_tab):
-        """5-up MetricsBar leaderboard — one column per model with MAPE."""
-        if active_tab != "tab-models":
-            return no_update
-        return _build_models_leaderboard(region)
-
-    @app.callback(
-        [
-            Output("tab3-metrics-table", "children"),
-            Output("tab3-residuals-time", "figure"),
-            Output("tab3-residuals-hist", "figure"),
-            Output("tab3-residuals-pred", "figure"),
-            Output("tab3-error-heatmap", "figure"),
-            Output("tab3-shap", "figure"),
-        ],
-        [
-            Input("demand-store", "data"),
-            Input("dashboard-tabs", "active_tab"),
-            Input("tab3-model-selector", "value"),
-        ],
-        [State("region-selector", "value")],
-        prevent_initial_call=True,
-    )
-    def update_models_tab(demand_json, active_tab, selected_models, region):
-        """Update Models tab diagnostics using model service."""
-        if active_tab != "tab-models":
-            return [no_update] * 6
-        if not demand_json:
-            empty = _empty_figure("Loading...")
-            return html.P("Loading..."), empty, empty, empty, empty, empty
-        if not selected_models:
-            empty = _empty_figure("Select at least one model to view diagnostics.")
-            return html.P("No model selected."), empty, empty, empty, empty, empty
-
-        # Redis fast path is valid only for ensemble-only diagnostics payloads.
-        if region:
-            redis_result = _models_tab_from_redis(region, selected_models)
-            if redis_result is not None:
-                return redis_result
-
-        # ── v1 compute fallback ─────────────────────────────
-        demand_df = pd.read_json(io.StringIO(demand_json))
-        demand_df["timestamp"] = pd.to_datetime(demand_df["timestamp"])
-        actual = demand_df["demand_mw"].values
-
-        from models.model_service import get_forecasts, get_model_metrics
-
-        forecasts = get_forecasts(region, demand_df, selected_models)
-        # Read metrics from the same Redis-first source the leaderboard
-        # uses (``get_model_metrics``) so the table and the leaderboard
-        # always show consistent numbers — even if ``get_forecasts``
-        # fell back to fresh simulated forecasts whose computed metrics
-        # would otherwise diverge from the Redis payload's metrics.
-        metrics = get_model_metrics(region) or forecasts.get("metrics", {})
-        model_order = ["prophet", "arima", "xgboost", "ensemble"]
-        selected = [m for m in model_order if m in set(selected_models)]
-
-        # Metrics table
-        name_map = {
-            "Prophet": "prophet",
-            "SARIMAX": "arima",
-            "XGBoost": "xgboost",
-            "Ensemble": "ensemble",
-        }
-        rows = []
-        for display_name, key in name_map.items():
-            if key not in selected:
-                continue
-            m = metrics.get(key, {})
-            rows.append(
-                html.Tr(
-                    [
-                        html.Td(display_name, style={"fontWeight": "600"}),
-                        html.Td(_format_metric(m, "mape", "{:.2f}%")),
-                        html.Td(_format_metric(m, "rmse", "{:.0f}")),
-                        html.Td(_format_metric(m, "mae", "{:.0f}")),
-                        html.Td(_format_metric(m, "r2", "{:.4f}")),
-                    ]
-                )
-            )
-        table = html.Table(
-            [
-                html.Thead(html.Tr([html.Th(h) for h in ["Model", "MAPE", "RMSE", "MAE", "R²"]])),
-                html.Tbody(rows),
-            ],
-            className="metrics-table",
-        )
-
-        timestamps = demand_df["timestamp"]
-        model_labels = {
-            "prophet": "Prophet",
-            "arima": "SARIMAX",
-            "xgboost": "XGBoost",
-            "ensemble": "Ensemble",
-        }
-        model_colors = {
-            "prophet": COLORS["prophet"],
-            "arima": COLORS["arima"],
-            "xgboost": COLORS["xgboost"],
-            "ensemble": COLORS["ensemble"],
-        }
-        model_residuals: dict[str, np.ndarray] = {}
-        model_predictions: dict[str, np.ndarray] = {}
-        for model_key in selected:
-            pred = forecasts.get(model_key)
-            if isinstance(pred, np.ndarray) and len(pred) == len(actual):
-                model_predictions[model_key] = pred
-                model_residuals[model_key] = actual - pred
-
-        if not model_residuals:
-            empty = _empty_figure("No residual diagnostics available for the selected model(s).")
-            return (
-                table,
-                empty,
-                empty,
-                empty,
-                empty,
-                _empty_figure("Select XGBoost to view SHAP feature importance."),
-            )
-
-        # uirevision keyed on region + sorted model selection so zoom/legend
-        # state survives data refresh but resets when the user picks new models.
-        uirev = f"{region}:{','.join(sorted(selected))}"
-
-        fig_resid_time = go.Figure()
-        for model_key, residuals in model_residuals.items():
-            fig_resid_time.add_trace(
-                go.Scatter(
-                    x=timestamps,
-                    y=residuals,
-                    mode="lines",
-                    name=model_labels.get(model_key, model_key.title()),
-                    line=dict(color=model_colors.get(model_key, COLORS["actual"]), width=1),
-                )
-            )
-        fig_resid_time.add_hline(y=0, line=dict(color="#F7FAFC", dash="dash", width=0.5))
-        fig_resid_time.update_layout(**_layout(uirevision=uirev, yaxis_title="Residual (MW)"))
-
-        fig_resid_hist = go.Figure()
-        for model_key, residuals in model_residuals.items():
-            fig_resid_hist.add_trace(
-                go.Histogram(
-                    x=residuals,
-                    nbinsx=50,
-                    name=model_labels.get(model_key, model_key.title()),
-                    marker_color=model_colors.get(model_key, COLORS["actual"]),
-                    opacity=0.6,
-                )
-            )
-        fig_resid_hist.update_layout(
-            **_layout(
-                uirevision=uirev,
-                barmode="overlay",
-                xaxis_title="Residual (MW)",
-                yaxis_title="Count",
-            )
-        )
-
-        fig_resid_pred = go.Figure()
-        for model_key, residuals in model_residuals.items():
-            preds = model_predictions[model_key]
-            fig_resid_pred.add_trace(
-                go.Scatter(
-                    x=preds,
-                    y=residuals,
-                    mode="markers",
-                    name=model_labels.get(model_key, model_key.title()),
-                    marker=dict(
-                        size=3,
-                        color=model_colors.get(model_key, COLORS["actual"]),
-                        opacity=0.35,
-                    ),
-                )
-            )
-        fig_resid_pred.add_hline(y=0, line=dict(color="#F7FAFC", dash="dash", width=0.5))
-        fig_resid_pred.update_layout(
-            **_layout(uirevision=uirev, xaxis_title="Predicted (MW)", yaxis_title="Residual (MW)")
-        )
-
-        hours_of_day = timestamps.dt.hour
-        fig_heatmap = go.Figure()
-        for model_key, residuals in model_residuals.items():
-            error_by_hour = pd.DataFrame({"hour": hours_of_day, "abs_error": np.abs(residuals)})
-            hourly_error = error_by_hour.groupby("hour")["abs_error"].mean()
-            fig_heatmap.add_trace(
-                go.Bar(
-                    x=hourly_error.index,
-                    y=hourly_error.values,
-                    name=model_labels.get(model_key, model_key.title()),
-                    marker_color=model_colors.get(model_key, COLORS["actual"]),
-                    opacity=0.85,
-                )
-            )
-        fig_heatmap.update_layout(**_layout(uirevision=uirev, barmode="group"))
-        fig_heatmap.update_layout(
-            **_layout(uirevision=uirev, xaxis_title="Hour of Day", yaxis_title="Mean |Error| (MW)")
-        )
-
-        if "xgboost" in selected:
-            feature_names, importance_vals = _get_feature_importance(region)
-            fig_shap = go.Figure(
-                go.Bar(
-                    x=importance_vals[::-1],
-                    y=feature_names[::-1],
-                    orientation="h",
-                    marker_color=COLORS["xgboost"],
-                )
-            )
-            fig_shap.update_layout(**_layout(uirevision=uirev, xaxis_title="Feature Importance"))
-        else:
-            fig_shap = _empty_figure("SHAP is available only for XGBoost. Select XGBoost above.")
-
-        return table, fig_resid_time, fig_resid_hist, fig_resid_pred, fig_heatmap, fig_shap
+    # ── 4. MODELS / DIAGNOSTICS TAB ────────────────────────────
+    # Step 10c of the register_callbacks split (issue #87) moves the
+    # three Models tab callbacks into ``_callbacks_models.py``. The
+    # section header was historically labeled "TAB 1: DEMAND FORECAST"
+    # — corrected here to match the actual tab the callbacks update.
+    register_models_callbacks(app)
 
     # ── 7. TAB 4: GENERATION & NET LOAD ──────────────────────
 
@@ -1460,597 +1222,24 @@ def register_callbacks(app):
 
         return new_mode, header_class, confidence_style, banner_style
 
-    # ── FORECAST TAB (R4a — v2 linear stack + inline panels) ────
-    # Hero chart + 4-up MetricsBar + InsightCard are still driven by
-    # ``update_demand_outlook`` below (existing 9-output callback,
-    # preserved). Small new callbacks fill the v2 title block, the
-    # ModelMetricsCard slot, and (R4a-2) the inline Drivers panel
-    # rendered when its collapse opens.
-
-    # 3 clientside toggles — flip is_open on each panel collapse.
-    # Generic JS could pattern-match, but explicit is clearer.
-    for _panel_key in ("drivers", "generation", "scenarios"):
-        app.clientside_callback(
-            "function(n, is_open) { return n ? !is_open : is_open; }",
-            Output(f"forecast-panel-{_panel_key}-collapse", "is_open"),
-            Input(f"forecast-panel-toggle-{_panel_key}", "n_clicks"),
-            State(f"forecast-panel-{_panel_key}-collapse", "is_open"),
-            prevent_initial_call=True,
-        )
-
-    @app.callback(
-        Output("forecast-drivers-content", "children"),
-        [
-            Input("forecast-panel-drivers-collapse", "is_open"),
-            Input("weather-store", "data"),
-        ],
-        State("dashboard-tabs", "active_tab"),
-    )
-    def update_forecast_drivers_panel(is_open, weather_json, active_tab):
-        """Render the 3-up Drivers KPI grid (Temperature / Wind / Solar).
-
-        Lazy: only computes when the collapse is open and the user is
-        on the Forecast tab (avoid spending render cost while collapsed).
-        """
-        if active_tab != "tab-outlook" or not is_open:
-            return no_update
-        return _build_drivers_panel(weather_json)
-
-    @app.callback(
-        Output("forecast-generation-content", "children"),
-        [
-            Input("forecast-panel-generation-collapse", "is_open"),
-            Input("region-selector", "value"),
-            Input("demand-store", "data"),
-        ],
-        State("dashboard-tabs", "active_tab"),
-    )
-    def update_forecast_generation_panel(is_open, region, demand_json, active_tab):
-        """Render the stacked-area fuel mix + 3-up sub-MetricsBar.
-
-        Lazy: only computes when the collapse is open and the user is
-        on the Forecast tab.
-        """
-        if active_tab != "tab-outlook" or not is_open:
-            return no_update
-        return _build_generation_panel(region, demand_json)
-
-    # Scenarios panel — preset chip click writes deltas into the 3 sliders.
-    @app.callback(
-        [
-            Output("forecast-scn-temp", "value"),
-            Output("forecast-scn-wind", "value"),
-            Output("forecast-scn-solar", "value"),
-        ],
-        Input({"type": "scenario-preset", "index": ALL}, "n_clicks"),
-        prevent_initial_call=True,
-    )
-    def apply_scenario_preset(_clicks):
-        """Apply a preset's temperature/wind/solar deltas to the three sliders."""
-        from components.tab_demand_outlook import _SCENARIO_PRESETS
-
-        triggered = ctx.triggered_id
-        if not isinstance(triggered, dict) or "index" not in triggered:
-            return no_update, no_update, no_update
-        # Ignore noop dispatches with all-zero clicks
-        if not any(c for c in (_clicks or []) if c):
-            return no_update, no_update, no_update
-        preset = _SCENARIO_PRESETS.get(triggered["index"])
-        if not preset:
-            return no_update, no_update, no_update
-        deltas = preset["deltas"]
-        return deltas["temp"], deltas["wind"], deltas["solar"]
-
-    # Slider readouts (clientside — instant, no Python round-trip)
-    for _key, _unit in (("temp", "°F"), ("wind", "mph"), ("solar", "W/m²")):
-        app.clientside_callback(
-            f"function(v) {{ if (v === null || v === undefined) return '0 {_unit}'; "
-            f"const sign = v > 0 ? '+' : ''; return sign + v + ' {_unit}'; }}",
-            Output(f"forecast-scn-{_key}-readout", "children"),
-            Input(f"forecast-scn-{_key}", "value"),
-        )
-
-    @app.callback(
-        [
-            Output("forecast-scenarios-kpis", "children"),
-            Output("forecast-scenarios-chart", "figure"),
-        ],
-        [
-            Input("forecast-panel-scenarios-collapse", "is_open"),
-            Input("forecast-scn-temp", "value"),
-            Input("forecast-scn-wind", "value"),
-            Input("forecast-scn-solar", "value"),
-            Input("region-selector", "value"),
-            Input("demand-store", "data"),
-        ],
-        State("dashboard-tabs", "active_tab"),
-    )
-    def update_forecast_scenarios_panel(
-        is_open, temp_d, wind_d, solar_d, region, demand_json, active_tab
-    ):
-        """Render the 4-up delta MetricsBar + baseline-vs-scenario chart.
-
-        Lazy: only fires when the collapse is open and the user is on the
-        Forecast tab. Uses a heuristic impact model (no full ensemble re-run)
-        so the slider feels instant; the existing Scenarios tab still hosts
-        the trained-model simulation for full-fidelity what-ifs.
-        """
-        if active_tab != "tab-outlook" or not is_open:
-            return no_update, no_update
-        return _build_scenarios_panel(temp_d, wind_d, solar_d, region, demand_json)
-
-    @app.callback(
-        Output("outlook-title", "children"),
-        [
-            Input("region-selector", "value"),
-            Input("dashboard-tabs", "active_tab"),
-        ],
-    )
-    def update_outlook_title(region, active_tab):
-        """Page-title block for the Forecast tab."""
-        if active_tab != "tab-outlook":
-            return no_update
-
-        region = region or "FPL"
-        region_name = REGION_NAMES.get(region, region)
-        return build_page_title(
-            "Forecast",
-            f"24h–30d demand outlook with confidence bands · {region_name}",
-        )
-
-    @app.callback(
-        Output("outlook-model-card", "children"),
-        [
-            Input("region-selector", "value"),
-            Input("outlook-model", "value"),
-            Input("dashboard-tabs", "active_tab"),
-        ],
-    )
-    def update_outlook_model_card(region, model_name, active_tab):
-        """Render the horizontal MAPE/RMSE/MAE/R² bar for the active model."""
-        if active_tab != "tab-outlook":
-            return no_update
-
-        try:
-            from models.model_service import get_model_metrics, is_trained
-        except ImportError:
-            return html.Div()
-
-        region = region or "FPL"
-        metrics_dict = get_model_metrics(region) or {}
-        if model_name not in metrics_dict:
-            # Fall back to any available model
-            if not metrics_dict:
-                return html.Div()
-            model_name = next(iter(metrics_dict.keys()))
-
-        m = metrics_dict[model_name]
-        formatted = {
-            "MAPE": f"{m.get('mape', 0.0):.1f}%",
-            "RMSE": f"{m.get('rmse', 0.0):,.0f} MW",
-            "MAE": f"{m.get('mae', 0.0):,.0f} MW",
-            "R²": f"{m.get('r2', 0.0):.3f}",
-        }
-        name = "XGBoost" if model_name == "xgboost" else model_name.title()
-        badge = "trained" if is_trained(region) else "simulated"
-        return build_model_metrics_card(model_name=name, metrics=formatted, badge=badge)
-
-    @app.callback(
-        [
-            Output("outlook-chart", "figure"),
-            Output("outlook-data-through", "children"),
-            Output("outlook-peak", "children"),
-            Output("outlook-peak-time", "children"),
-            Output("outlook-avg", "children"),
-            Output("outlook-min", "children"),
-            Output("outlook-min-time", "children"),
-            Output("outlook-range", "children"),
-            Output("tab2-insight-card", "children"),
-        ],
-        [
-            Input("outlook-horizon", "value"),
-            Input("outlook-model", "value"),
-            Input("dashboard-tabs", "active_tab"),
-            Input("demand-store", "data"),
-            Input("persona-selector", "value"),
-        ],
-        [
-            State("weather-store", "data"),
-            State("region-selector", "value"),
-        ],
-        prevent_initial_call=True,
-    )
-    def update_demand_outlook(
-        horizon,
-        model_name,
-        active_tab,
-        demand_json,
-        persona_id,
-        weather_json,
-        region,
-    ):
-        """Generate forward-looking demand forecast."""
-        # Only run when this tab is active — avoids 10s+ model training on page load
-        if active_tab != "tab-outlook":
-            return [no_update] * 9
-
-        log.info("outlook_callback_start", horizon=horizon, model=model_name, region=region)
-
-        horizon_hours = int(horizon)
-        empty_insight = html.Div()
-
-        # ── v2 Redis fast path ──────────────────────────────
-        if region:
-            redis_result = _outlook_tab_from_redis(
-                region, horizon_hours, model_name, demand_json, weather_json, persona_id
-            )
-            if redis_result is not None:
-                return redis_result
-
-        # uirevision keyed on region + horizon so zoom/legend state persists
-        # across data refresh but resets when the user picks a new horizon.
-        uirev = f"{region}:{horizon_hours}"
-
-        # ── v1 compute fallback ─────────────────────────────
-        if not demand_json or not weather_json:
-            fig = go.Figure()
-            fig.update_layout(**_layout(uirevision=uirev))
-            fig.add_annotation(
-                text="Loading data...", xref="paper", yref="paper", x=0.5, y=0.5, showarrow=False
-            )
-            return (
-                fig,
-                "Loading...",
-                "Loading...",
-                "",
-                "Loading...",
-                "Loading...",
-                "",
-                "Loading...",
-                empty_insight,
-            )
-
-        try:
-            demand_df = pd.read_json(io.StringIO(demand_json))
-            weather_df = pd.read_json(io.StringIO(weather_json))
-        except Exception as e:
-            log.error("outlook_parse_error", error=str(e))
-            fig = go.Figure()
-            fig.update_layout(**_layout(uirevision=uirev))
-            return (
-                fig,
-                "Error",
-                "No data",
-                "",
-                "No data",
-                "No data",
-                "",
-                "No data",
-                empty_insight,
-            )
-
-        # Get the data through date (last timestamp in demand data)
-        demand_df["timestamp"] = pd.to_datetime(demand_df["timestamp"])
-        data_through = demand_df["timestamp"].max()
-        data_through_str = data_through.strftime("%Y-%m-%d %H:%M UTC")
-
-        # Run the forecast
-        result = _run_forecast_outlook(demand_df, weather_df, horizon_hours, model_name, region)
-
-        if "error" in result:
-            # Soften the warming case (pipeline still populating Redis after
-            # a deploy / cache eviction) — that's an expected transient state,
-            # not a hard failure. Keep the loud message for genuine errors.
-            is_warming = result["error"] == "warming"
-            text = (
-                "Pipeline is warming up — forecast will appear shortly"
-                if is_warming
-                else f"Forecast failed: {result['error']}"
-            )
-            color = "#71717a" if is_warming else "#f87171"  # tertiary | danger
-            fig = go.Figure()
-            fig.update_layout(**_layout(uirevision=uirev))
-            fig.add_annotation(
-                text=text,
-                xref="paper",
-                yref="paper",
-                x=0.5,
-                y=0.5,
-                showarrow=False,
-                font=dict(color=color, size=14),
-            )
-            return (
-                fig,
-                data_through_str,
-                "No data",
-                "",
-                "No data",
-                "No data",
-                "",
-                "No data",
-                empty_insight,
-            )
-
-        timestamps = pd.to_datetime(result["timestamps"])
-        predictions = result["predictions"]
-
-        # Build per-point tooltip strings (NEXD-13)
-        tooltips = None
-        try:
-            from config import feature_enabled
-
-            if feature_enabled("inline_tooltips"):
-                from data.explainability import build_tooltip_strings
-
-                shap_data = result.get("shap_data")
-                tooltips = build_tooltip_strings(
-                    shap_values=shap_data.get("shap_values") if shap_data else None,
-                    feature_names=shap_data.get("feature_names") if shap_data else None,
-                    model_dict=result.get("model_dict"),
-                    n_points=len(predictions),
-                    model_name=model_name,
-                )
-        except Exception:
-            log.debug("tooltip_build_skipped")
-
-        # Calculate KPIs
-        peak_val = np.max(predictions)
-        peak_idx = np.argmax(predictions)
-        peak_time = timestamps[peak_idx].strftime("%a %H:%M")
-
-        min_val = np.min(predictions)
-        min_idx = np.argmin(predictions)
-        min_time = timestamps[min_idx].strftime("%a %H:%M")
-
-        avg_val = np.mean(predictions)
-        range_val = peak_val - min_val
-
-        # Build chart
-        fig = go.Figure()
-
-        # Forecast line (model-aware color + dash pattern)
-        model_style = LINE_STYLES.get(
-            model_name, {"color": COLORS["ensemble"], "width": 2, "dash": "solid"}
-        )
-        # Forecast trace with optional SHAP tooltips (NEXD-13)
-        forecast_kwargs: dict = dict(
-            x=timestamps,
-            y=predictions,
-            mode="lines",
-            name=f"{model_name.upper()} Forecast",
-            line=dict(
-                color=COLORS.get(model_name, COLORS["ensemble"]),
-                width=model_style.get("width", 2),
-                dash=model_style.get("dash", "solid"),
-            ),
-            fill="tozeroy",
-            fillcolor="rgba(56,208,255,0.10)",
-        )
-        if tooltips and any(tooltips):
-            forecast_kwargs["customdata"] = tooltips
-            forecast_kwargs["hovertemplate"] = (
-                "<b>%{x|%a %b %d %H:%M}</b><br>Demand: %{y:,.0f} MW<br>%{customdata}<extra></extra>"
-            )
-        fig.add_trace(go.Scatter(**forecast_kwargs))
-
-        # Add peak marker
-        fig.add_trace(
-            go.Scatter(
-                x=[timestamps[peak_idx]],
-                y=[peak_val],
-                mode="markers+text",
-                name="Peak",
-                marker=dict(color="#FF5C7A", size=12, symbol="triangle-up"),
-                text=[f"Peak: {peak_val:,.0f} MW"],
-                textposition="top center",
-                showlegend=False,
-            )
-        )
-        # Add min marker
-        fig.add_trace(
-            go.Scatter(
-                x=[timestamps[min_idx]],
-                y=[min_val],
-                mode="markers+text",
-                name="Min",
-                marker=dict(color="#3b82f6", size=10, symbol="triangle-down"),
-                text=[f"Min: {min_val:,.0f} MW"],
-                textposition="bottom center",
-                showlegend=False,
-            )
-        )
-        interval_meta = _add_confidence_bands(
-            fig, timestamps, predictions, horizon_hours, region=region, model_name=model_name
-        )
-        _add_trailing_actuals(fig, demand_json)
-
-        # Layout
-        horizon_labels = {24: "24-Hour", 168: "7-Day", 720: "30-Day"}
-        interval_caption = ""
-        if interval_meta.get("method") == "empirical":
-            interval_caption = (
-                f"<br><sup>80% empirical prediction interval "
-                f"(calibration window: last "
-                f"{int(interval_meta.get('calibration_window_hours', 0))}h)</sup>"
-            )
-        fig.update_layout(
-            **_layout(
-                uirevision=uirev,
-                title=(
-                    f"{horizon_labels.get(horizon_hours, '')} {model_name.upper()} Demand Forecast — {region}"
-                    f"{interval_caption}"
-                ),
-                xaxis_title="Date/Time",
-                yaxis_title="Demand (MW)",
-                hovermode="x unified",
-            )
-        )
-
-        # Format KPI strings
-        peak_str = f"{peak_val:,.0f} MW"
-        avg_str = f"{avg_val:,.0f} MW"
-        min_str = f"{min_val:,.0f} MW"
-        range_str = f"{range_val:,.0f} MW"
-
-        # Generate insights
-        from components.insights import build_insight_card, generate_tab2_insights
-
-        persona = persona_id or "grid_ops"
-        tab2_insights = generate_tab2_insights(
-            persona,
-            region or "FPL",
-            predictions,
-            timestamps,
-            model_name=model_name,
-            horizon_hours=horizon_hours,
-            weather_df=weather_df,
-        )
-        insight_card = build_insight_card(tab2_insights, persona, "tab-outlook")
-
-        log.info("outlook_callback_complete", horizon=horizon_hours, peak=peak_str)
-        return (
-            fig,
-            data_through_str,
-            peak_str,
-            peak_time,
-            avg_str,
-            min_str,
-            min_time,
-            range_str,
-            insight_card,
-        )
-
-    # ── FORECAST REPLAY SELECTOR (NEXD-14) ──────────────────────
-
-    @app.callback(
-        [
-            Output("replay-selector", "options"),
-            Output("replay-selector", "value"),
-            Output("replay-container", "style"),
-        ],
-        [
-            Input("outlook-horizon", "value"),
-            Input("outlook-model", "value"),
-            Input("dashboard-tabs", "active_tab"),
-            Input("region-selector", "value"),
-        ],
-        prevent_initial_call=True,
-    )
-    def populate_replay_selector(horizon, model_name, active_tab, region):
-        """Populate the replay dropdown with available forecast snapshots."""
-        from config import feature_enabled
-
-        hidden = {"display": "none"}
-        default_opts = [{"label": "Current", "value": "current"}]
-
-        if active_tab != "tab-outlook" or not feature_enabled("forecast_replay"):
-            return default_opts, "current", hidden
-
-        try:
-            from data.forecast_history import build_replay_options
-
-            horizon_hours = int(horizon) if horizon else 168
-            options = build_replay_options(region or "FPL", horizon_hours, model_name or "xgboost")
-            # Hide if only "Current" (no historical snapshots to compare)
-            visible = {"display": "block"} if len(options) > 1 else hidden
-            return options, "current", visible
-        except Exception:
-            log.debug("replay_selector_populate_failed")
-            return default_opts, "current", hidden
-
-    # ── FORECAST REPLAY OVERLAY (NEXD-14) ───────────────────────
-
-    @app.callback(
-        [
-            Output("outlook-chart", "figure", allow_duplicate=True),
-            Output("replay-label", "children"),
-        ],
-        [Input("replay-selector", "value")],
-        [
-            State("outlook-chart", "figure"),
-            State("outlook-horizon", "value"),
-            State("outlook-model", "value"),
-            State("region-selector", "value"),
-        ],
-        prevent_initial_call=True,
-    )
-    def overlay_replay_snapshot(replay_value, current_fig, horizon, model_name, region):
-        """Overlay a historical forecast snapshot on the current chart.
-
-        This is a lightweight callback — it only reads from SQLite and
-        patches the existing figure.  It never recomputes a forecast.
-        """
-        if current_fig is None:
-            return no_update, ""
-
-        # Fast path: if "current" and no replay traces exist, skip figure round-trip
-        has_replay_traces = any(
-            (t.get("name") or "").startswith("Forecast from ")
-            for t in (current_fig.get("data") or [])
-        )
-        if (not replay_value or replay_value == "current") and not has_replay_traces:
-            return no_update, ""
-
-        fig = go.Figure(current_fig)
-
-        # Strip any previously added replay traces
-        fig.data = [t for t in fig.data if not (t.name or "").startswith("Forecast from ")]
-
-        if not replay_value or replay_value == "current":
-            return fig, ""
-
-        try:
-            from config import feature_enabled
-
-            if not feature_enabled("forecast_replay"):
-                return fig, ""
-
-            from data.forecast_history import get_forecast_snapshot
-
-            horizon_hours = int(horizon) if horizon else 168
-            snap = get_forecast_snapshot(
-                region or "FPL", horizon_hours, model_name or "xgboost", replay_value
-            )
-            if snap:
-                from datetime import datetime as _dt
-
-                try:
-                    snap_label = _dt.fromisoformat(snap["scored_at"]).strftime("%b %d %H:%M UTC")
-                except (ValueError, TypeError):
-                    snap_label = snap["scored_at"][:16]
-
-                fig.add_trace(
-                    go.Scatter(
-                        x=pd.to_datetime(snap["timestamps"]),
-                        y=snap["predictions"],
-                        mode="lines",
-                        name=f"Forecast from {snap_label}",
-                        line=dict(color="#A8B3C7", width=2, dash="dash"),
-                        opacity=0.6,
-                    )
-                )
-                return fig, f"Comparing with forecast from {snap_label}"
-        except Exception:
-            log.debug("replay_overlay_failed")
-
-        return fig, ""
+    # ── FORECAST TAB (R4a + Replay) ─────────────────────────────
+    # Step 10c of the register_callbacks split (issue #87) moves the
+    # 9 Forecast tab callbacks (3 panel toggles + 3 inline panels +
+    # scenario presets + title + model card + the big 9-output
+    # ``update_demand_outlook`` + 2 Forecast Replay callbacks) into
+    # ``_callbacks_forecast.py``.
+    register_forecast_callbacks(app)
 
 
 # ── HELPER FUNCTIONS ──────────────────────────────────────────
-
-
-# ── Defensive readers ──────────────────────────────────────────
 #
-# EIA-930 has a publishing lag for the most recent hour, especially for
-# newer / smaller BAs (PSCO, NEVP, AZPS observed 2026-05-02). Until the
-# row catches up the demand series ends with NaN — sometimes preceded by
-# zero (EIA's other "missing observation" marker). Anywhere we read the
-# most recent value off a demand list or Series we route through this
-# helper so downstream code never has to ``isnan`` or guard ``prev / 0``.
-#
-# Companion to the ``> 0`` filter in ``_build_overview_metrics_items`` —
-# this formalizes that pattern as the single source of truth, so future
-# tabs (US Grid, Risk, Models) can't accidentally re-introduce the bug.
+# Per the callbacks.py decomposition (issue #87) every tab-specific
+# helper now lives in its corresponding ``_callbacks_<tab>.py`` module.
+# Cross-cutting utilities (caches, layout, prediction intervals,
+# defensive readers) live in ``_callbacks_shared.py``. The explicit
+# re-imports at the top of this file expose every helper under the
+# ``components.callbacks`` namespace so existing import sites
+# (``from components.callbacks import <X>``) continue to resolve.
 
 
 # ── V1.β: US GRID TAB HELPERS ────────────────────────────────
